requirements.txt
# requirements.txt - Dashboard Dependencies
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
google-cloud-bigquery>=3.15.0
pandas-gbq>=0.20.0
google-auth>=2.23.0
google-generativeai>=0.4.0
